Bruh
I hate what Pogba did to Poland, this is some GFM level event gay shit.
Anyone with half a braincell knows events suck ass, its impossible to make sense of this fucking tree. You need a fucking excel spreadsheet to understand what the fuck is even happening. Like there is straight up no explanation too, in literally every mp game the Poland player just gets stuck with a Silean and Hungarian Puppet and his starting land and cant do shit the entire game because half his country is owned by the ai. Like fucking hell is it really so hard to just make 5 decisions that have an explanation in their description if they are mutually exclusive? Or atleast give a starting event that explains the situation and the paths. This shit is so fucking retarded, whoever made this should kys.